Fucked bug !

+++ dlls/extdll.h
@@ -43,7 +43,7 @@
 #else // _WIN32
 #define FALSE 0
 #define TRUE (!FALSE)
-typedef unsigned int ULONG;
+typedef unsigned long ULONG;
 typedef unsigned char BYTE;
 typedef int BOOL;
 #define MAX_PATH PATH_MAX
